test(opencode): isolate MCP reconnect coverage
